// packages/shared/src/enums/ReflectorType.ts
// 직각의 위치 기준으로 명명
// TopLeft: 직각이 왼쪽 위에 위치
// TopRight: 직각이 오른쪽 위에 위치
// BottomLeft: 직각이 왼쪽 아래에 위치
// BottomRight: 직각이 오른쪽 아래에 위치
export enum ReflectorType {
  None = 0,
  TopLeft = 1,
  TopRight = 2,
  BottomLeft = 3,
  BottomRight = 4,
}